MI PROGRAMA FUNCIONA SOLO FALTA  IMPLEMENTARLA CON LA  CLASE PUNTO

+++ ConsoleApp4/Punto.cs
@@ -8,7 +8,7 @@ namespace ConsoleApp4
 {
     public class Punto
     {
-        public int x { get; set; }
-        public int y { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
     }
 }
